+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## DEV
 
 * Improved handling JSON parsing error
+* Improved handling 403 error in stores API
 
 ## 0.1.0 (2019-02-15)